Consolidates test in the /tests directory.
Moves addressing unit tests into /tests folder
