+++ tests/addressing/role_members.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 # -----------------------------------------------------------------------------
 
+import pytest
 import unittest
 import logging
 from uuid import uuid4
@@ -22,17 +23,19 @@ from rbac_addressing.addresser import AddressSpace
 LOGGER = logging.getLogger(__name__)
 
 
-class TestTaskAdminsAddresser(unittest.TestCase):
-    def test_determine_task_admin_addr(self):
-        """Tests that a specific task_id and admin_id generates the
-        expected task admin address, and thus is probably deterministic.
+@pytest.mark.unit
+@pytest.mark.addressing
+class TestRoleMembersAddresser(unittest.TestCase):
+    def test_determine_role_member_addr(self):
+        """Tests that a specific role_id and member_id generates the
+        expected role member address, and thus is probably deterministic.
         """
 
-        task_id = "99968acb8f1a48b3a4bc21e2cd252e67"
-        admin_id = "966ab67317234df489adb4bc1f517b88"
-        expected_address = "9f44481e326a1713a905b26359fc8d\
-a2817c1a5f67de6f464701f0c10042da345d2848"
-        address = addresser.make_task_admins_address(task_id, admin_id)
+        role_id = "99968acb8f1a48b3a4bc21e2cd252e67"
+        member_id = "966ab67317234df489adb4bc1f517b88"
+        expected_address = "9f444809326a1713a905b26359fc8d\
+a2817c1a5f67de6f464701f0c10042da345d2833"
+        address = addresser.make_role_members_address(role_id, member_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -57,18 +60,18 @@ a2817c1a5f67de6f464701f0c10042da345d2848"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.TASKS_ADMINS,
-            "The address created must be a Task Attributes address.",
+            AddressSpace.ROLES_MEMBERS,
+            "The address created must be a Role Attributes address.",
         )
 
-    def test_generated_task_admin_addr(self):
-        """Tests the task admin address creation function as well as the
+    def test_generated_role_member_addr(self):
+        """Tests the role member address creation function as well as the
         address_is function.
         """
 
-        task_id = uuid4().hex
-        admin_id = uuid4().hex
-        address = addresser.make_task_admins_address(task_id, admin_id)
+        role_id = uuid4().hex
+        member_id = uuid4().hex
+        address = addresser.make_role_members_address(role_id, member_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -89,6 +92,6 @@ a2817c1a5f67de6f464701f0c10042da345d2848"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.TASKS_ADMINS,
-            "The address created must be a Task Attributes address.",
+            AddressSpace.ROLES_MEMBERS,
+            "The address created must be a Role Attributes address.",
         )
