+++ tests/addressing/role_tasks.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 # -----------------------------------------------------------------------------
 
+import pytest
 import unittest
 import logging
 from uuid import uuid4
@@ -22,16 +23,19 @@ from rbac_addressing.addresser import AddressSpace
 LOGGER = logging.getLogger(__name__)
 
 
-class TestRoleAttributesAddresser(unittest.TestCase):
-    def test_deterministic_role_address(self):
-        """Tests that a specific role_id generates the expected
-        role address, and thus is probably deterministic.
+@pytest.mark.unit
+@pytest.mark.addressing
+class TestRoleTasksAddresser(unittest.TestCase):
+    def test_determine_role_task_addr(self):
+        """Tests that a specific role_id and task_id generates the
+        expected role task address, and thus is probably deterministic.
         """
 
-        ident = "99968acb8f1a48b3a4bc21e2cd252e67"
+        role_id = "99968acb8f1a48b3a4bc21e2cd252e67"
+        task_id = "966ab67317234df489adb4bc1f517b88"
         expected_address = "9f444809326a1713a905b26359fc8d\
-a2817c1a5f67de6f464701f0c10042da345d2800"
-        address = addresser.make_role_attributes_address(ident)
+a2817c1a5f67de6f464701f0c10042da345d28c5"
+        address = addresser.make_role_tasks_address(role_id, task_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -56,17 +60,18 @@ a2817c1a5f67de6f464701f0c10042da345d2800"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.ROLES_ATTRIBUTES,
+            AddressSpace.ROLES_TASKS,
             "The address created must be a Role Attributes address.",
         )
 
-    def test_generated_role_address(self):
-        """Tests the role address creation function as well as the
+    def test_generated_role_task_addr(self):
+        """Tests the role task address creation function as well as the
         address_is function.
         """
 
-        ident = uuid4().hex
-        address = addresser.make_role_attributes_address(ident)
+        role_id = uuid4().hex
+        task_id = uuid4().hex
+        address = addresser.make_role_tasks_address(role_id, task_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -87,6 +92,6 @@ a2817c1a5f67de6f464701f0c10042da345d2800"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.ROLES_ATTRIBUTES,
+            AddressSpace.ROLES_TASKS,
             "The address created must be a Role Attributes address.",
         )
