+++ tests/addressing/sysadmin_owners.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 # -----------------------------------------------------------------------------
 
+import pytest
 import unittest
 import logging
 from uuid import uuid4
@@ -22,16 +23,18 @@ from rbac_addressing.addresser import AddressSpace
 LOGGER = logging.getLogger(__name__)
 
 
-class TestSysAdminAdminsAddresser(unittest.TestCase):
-    def test_det_sysadmin_admin_addr(self):
-        """Tests that a specific admin_id generates the expected
-        sysadmin admin address, and thus is probably deterministic.
+@pytest.mark.unit
+@pytest.mark.addressing
+class TestSysAdminOwnersAddresser(unittest.TestCase):
+    def test_det_sysadmin_owner_addr(self):
+        """Tests that a specific owner_id generates the expected
+        sysadmin owner address, and thus is probably deterministic.
         """
 
-        admin_id = "966ab67317234df489adb4bc1f517b88"
+        owner_id = "966ab67317234df489adb4bc1f517b88"
         expected_address = "9f4448000000000000000000000000\
-00000000000000000000000000000000000000f7"
-        address = addresser.make_sysadmin_admins_address(admin_id)
+00000000000000000000000000000000000000de"
+        address = addresser.make_sysadmin_owners_address(owner_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -56,17 +59,17 @@ class TestSysAdminAdminsAddresser(unittest.TestCase):
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.SYSADMIN_ADMINS,
+            AddressSpace.SYSADMIN_OWNERS,
             "The address created must be a SysAdmin Attributes address.",
         )
 
-    def test_gen_sysadmin_admin_addr(self):
-        """Tests the sysadmin admin address creation function as well as the
+    def test_gen_sysadmin_owner_addr(self):
+        """Tests the sysadmin owner address creation function as well as the
         address_is function.
         """
 
-        admin_id = uuid4().hex
-        address = addresser.make_sysadmin_admins_address(admin_id)
+        owner_id = uuid4().hex
+        address = addresser.make_sysadmin_owners_address(owner_id)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -87,6 +90,6 @@ class TestSysAdminAdminsAddresser(unittest.TestCase):
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.SYSADMIN_ADMINS,
+            AddressSpace.SYSADMIN_OWNERS,
             "The address created must be a SysAdmin Attributes address.",
         )
