+++ tests/addressing/task.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 # -----------------------------------------------------------------------------
 
+import pytest
 import unittest
 import logging
 from uuid import uuid4
@@ -22,17 +23,18 @@ from rbac_addressing.addresser import AddressSpace
 LOGGER = logging.getLogger(__name__)
 
 
-class TestRoleMembersAddresser(unittest.TestCase):
-    def test_determine_role_member_addr(self):
-        """Tests that a specific role_id and member_id generates the
-        expected role member address, and thus is probably deterministic.
+@pytest.mark.unit
+@pytest.mark.addressing
+class TestTaskAttributesAddresser(unittest.TestCase):
+    def test_deterministic_task_address(self):
+        """Tests that a specific task_id generates the expected
+        task address, and thus is probably deterministic.
         """
 
-        role_id = "99968acb8f1a48b3a4bc21e2cd252e67"
-        member_id = "966ab67317234df489adb4bc1f517b88"
-        expected_address = "9f444809326a1713a905b26359fc8d\
-a2817c1a5f67de6f464701f0c10042da345d2833"
-        address = addresser.make_role_members_address(role_id, member_id)
+        ident = "99968acb8f1a48b3a4bc21e2cd252e67"
+        expected_address = "9f44481e326a1713a905b26359fc8d\
+a2817c1a5f67de6f464701f0c10042da345d2800"
+        address = addresser.make_task_attributes_address(ident)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -57,18 +59,17 @@ a2817c1a5f67de6f464701f0c10042da345d2833"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.ROLES_MEMBERS,
-            "The address created must be a Role Attributes address.",
+            AddressSpace.TASKS_ATTRIBUTES,
+            "The address created must be a Task Attributes address.",
         )
 
-    def test_generated_role_member_addr(self):
-        """Tests the role member address creation function as well as the
+    def test_generated_task_address(self):
+        """Tests the task address creation function as well as the
         address_is function.
         """
 
-        role_id = uuid4().hex
-        member_id = uuid4().hex
-        address = addresser.make_role_members_address(role_id, member_id)
+        ident = uuid4().hex
+        address = addresser.make_task_attributes_address(ident)
 
         self.assertEqual(
             len(address), addresser.ADDRESS_LENGTH, "The address is 70 characters"
@@ -89,6 +90,6 @@ a2817c1a5f67de6f464701f0c10042da345d2833"
 
         self.assertEqual(
             addresser.address_is(address),
-            AddressSpace.ROLES_MEMBERS,
-            "The address created must be a Role Attributes address.",
+            AddressSpace.TASKS_ATTRIBUTES,
+            "The address created must be a Task Attributes address.",
         )
